Bring in the RX event bus and create an event
This is an event bus I created to initially turn into a library or
something, but it wasn't that complex so I just copied the java file
here. Not in use in any of the commited code yet. I have a large bundle
of uncommited things that are a start towards trying the MVP pattern,
and currently it looks ugly.