Refactoring and work on using an MVP strategy
A fair bit of renaming etc. to serparate data from utilities that help
with the data. Moving to the MVP style seems nice, but when also trying
to keep as much as possible clean of Android dependencies, its tricky.
Since I need context for this like sharedpreferences etc, can I actually
keep android stuff out of the model classes?