Reorganize things and get a better handle on dagger
Move the interactors to a separate package and module to allow reuse,
and just make things a little cleaner. Have the fragment be the thing
which creates the object graph, leaving the activity out of it, and
making it easier to pass the correct view in to the module. This should
also make it easier to have multiple fragments on screen with seperate
object graphs.

Still not sure the best way to get activity context though. Perhaps a
base activity as well, that just has a module providing context.