Create skeleton things for details view
Just make all the files necessary to create the details view, and
 first draft for the view/presenter interfaces. Begin to move the
retrofit stuff to a module, to allow for injection and moving away from
static things. For other classes, like the filehandler.class and utility
classes, I'm not sure the best way to adapt them to be nonstatic, for
dagger usage.