More into dagger modules and de-static some classes
To make things a little simpler, I've changed the retrofit client's
class to be instanitiable. IT was all final static to help it be a
singleton since apparently the restapdater was heavy to make. But it's
now provided as a singleton per activity using dagger. And given how android gets rid
of things or not, who knows if it was actually a single instance for the
life of the application before anyway.