Move the add locations dialog to mvp layout
Using dagger here makes it easier/nicer? to not use static classes for the
location provider. I still need to think about how that should work, as
far as which thing sits and displays while getting gps coordinates. It
would be easy to just push the request back to the list
fragment/presenter, but then how to indicate that its looking?